Correct link to Guzzle docs and add version
Guzzle 3.x is now deprecated, 6.x is the current version. There are significant code changes and PAW creates code for Guzzle 3.x. The link pointed to the latest documentation which is 6.x. Corrected link and added the version so users of the extension know which version of Guzzle code this creates.